Changed the navbar position
In the first stage, I aligned the navbar to the top, as on the LinkedIn site, but I moved it to the left for a more innovative design. At the current stage, when the mouse hovers over it, the text opens, but when it is not on it, I will make it a window where only the icons are visible.